requirements.txt
torch
torchvision
tqdm
numpy
pillow
pyyaml
streamlit
nltk
